+++ ricaun.Revit.ExtensibleJson.Tests/Utils/OneTimeOpenDocumentTest.cs
@@ -1,0 +1,35 @@
+﻿using Autodesk.Revit.ApplicationServices;
+using Autodesk.Revit.DB;
+using NUnit.Framework;
+
+namespace ricaun.Revit.ExtensibleJson.Tests.Utils
+{
+    /// <summary>
+    /// OneTimeOpenDocumentTest
+    /// </summary>
+    public class OneTimeOpenDocumentTest
+    {
+        protected Document document;
+        protected Application application;
+        protected virtual string FileName => null;
+
+        [OneTimeSetUp]
+        public void NewProjectDocument(Application application)
+        {
+            this.application = application;
+            if (string.IsNullOrEmpty(FileName))
+            {
+                document = application.NewProjectDocument(UnitSystem.Metric);
+                return;
+            }
+            document = application.OpenDocumentFile(FileName);
+        }
+
+        [OneTimeTearDown]
+        public void CloseProjectDocument()
+        {
+            document.Close(false);
+            document.Dispose();
+        }
+    }
+}