+++ ricaun.Revit.ExtensibleJson/IJsonService.cs
@@ -3,7 +3,7 @@
 namespace ricaun.Revit.ExtensibleJson
 {
     /// <summary>
-    /// IJsonService
+    /// Provides methods for serializing and deserializing JSON data.
     /// </summary>
     public interface IJsonService : IJsonService<object>
     {
@@ -11,45 +11,55 @@ namespace ricaun.Revit.ExtensibleJson
     }
 
     /// <summary>
-    /// IJsonService
+    /// Defines a service for JSON serialization and deserialization for a specific type.
     /// </summary>
-    /// <typeparam name="TJson"></typeparam>
+    /// <typeparam name="TJson">The type to serialize or deserialize.</typeparam>
     public interface IJsonService<TJson>
     {
         /// <summary>
-        /// GetSettings
+        /// Gets the <see cref="JsonSerializerSettings"/> used for serialization and deserialization.
         /// </summary>
-        /// <returns></returns>
+        /// <returns>
+        /// The <see cref="JsonSerializerSettings"/> instance.
+        /// </returns>
         JsonSerializerSettings GetSettings();
 
         /// <summary>
-        /// Serialize
+        /// Serializes the specified value to a JSON string.
         /// </summary>
-        /// <param name="value"></param>
-        /// <returns></returns>
+        /// <param name="value">The value to serialize.</param>
+        /// <returns>
+        /// A JSON string representation of the value.
+        /// </returns>
         string Serialize(TJson value);
 
         /// <summary>
-        /// SerializeObject
+        /// Serializes the specified object to a JSON string.
         /// </summary>
-        /// <typeparam name="T"></typeparam>
-        /// <param name="value"></param>
-        /// <returns></returns>
+        /// <typeparam name="T">The type of the object to serialize.</typeparam>
+        /// <param name="value">The object to serialize.</param>
+        /// <returns>
+        /// A JSON string representation of the object.
+        /// </returns>
         string SerializeObject<T>(T value);
 
         /// <summary>
-        /// Deserialize
+        /// Deserializes the specified JSON string to an object of type <typeparamref name="TJson"/>.
         /// </summary>
-        /// <param name="value"></param>
-        /// <returns></returns>
+        /// <param name="value">The JSON string to deserialize.</param>
+        /// <returns>
+        /// An object of type <typeparamref name="TJson"/> deserialized from the JSON string.
+        /// </returns>
         TJson Deserialize(string value);
 
         /// <summary>
-        /// DeserializeObject
+        /// Deserializes the specified JSON string to an object of type <typeparamref name="T"/>.
         /// </summary>
-        /// <typeparam name="T"></typeparam>
-        /// <param name="value"></param>
-        /// <returns></returns>
+        /// <typeparam name="T">The type of the object to deserialize.</typeparam>
+        /// <param name="value">The JSON string to deserialize.</param>
+        /// <returns>
+        /// An object of type <typeparamref name="T"/> deserialized from the JSON string.
+        /// </returns>
         T DeserializeObject<T>(string value);
     }
 }