+++ ricaun.Revit.ExtensibleJson/JsonService.cs
@@ -4,22 +4,26 @@ using ricaun.Revit.ExtensibleJson.Converters;
 namespace ricaun.Revit.ExtensibleJson
 {
     /// <summary>
-    /// JsonService
+    /// Provides JSON serialization and deserialization services using default settings and custom converters.
     /// </summary>
+    /// <remarks>
+    /// This class is a non-generic implementation of <see cref="JsonService{TJson}"/> for <see cref="object"/> type.
+    /// It implements <see cref="IJsonService"/> for convenience.
+    /// </remarks>
     public class JsonService : JsonService<object>, IJsonService
     {
 
     }
 
     /// <summary>
-    /// JsonService
+    /// Provides JSON serialization and deserialization services for a specific type using custom converters.
     /// </summary>
-    /// <typeparam name="TJson"></typeparam>
+    /// <typeparam name="TJson">The type to serialize and deserialize.</typeparam>
     public class JsonService<TJson> : IJsonService<TJson>
     {
         private readonly JsonSerializerSettings settings;
         /// <summary>
-        /// JsonService
+        /// Initializes a new instance of the <see cref="JsonService{TJson}"/> class with default settings and custom converters.
         /// </summary>
         public JsonService()
         {
@@ -29,48 +33,48 @@ namespace ricaun.Revit.ExtensibleJson
         }
 
         /// <summary>
-        /// Get JsonSerializerSettings
+        /// Gets the <see cref="JsonSerializerSettings"/> used for serialization and deserialization.
         /// </summary>
-        /// <returns></returns>
+        /// <returns>The <see cref="JsonSerializerSettings"/> instance.</returns>
         public JsonSerializerSettings GetSettings() => settings;
 
         /// <summary>
-        /// Serialize
+        /// Serializes the specified value to a JSON string.
         /// </summary>
-        /// <param name="value"></param>
-        /// <returns></returns>
+        /// <param name="value">The value to serialize.</param>
+        /// <returns>A JSON string representation of the value.</returns>
         public string Serialize(TJson value)
         {
             return SerializeObject<TJson>(value);
         }
 
         /// <summary>
-        /// SerializeObject
+        /// Serializes the specified value of type <typeparamref name="T"/> to a JSON string.
         /// </summary>
-        /// <typeparam name="T"></typeparam>
-        /// <param name="value"></param>
-        /// <returns></returns>
+        /// <typeparam name="T">The type of the value to serialize.</typeparam>
+        /// <param name="value">The value to serialize.</param>
+        /// <returns>A JSON string representation of the value.</returns>
         public string SerializeObject<T>(T value)
         {
             return JsonConvert.SerializeObject(value, settings);
         }
 
         /// <summary>
-        /// Deserialize
+        /// Deserializes the specified JSON string to an object of type <typeparamref name="TJson"/>.
         /// </summary>
-        /// <param name="value"></param>
-        /// <returns></returns>
+        /// <param name="value">The JSON string to deserialize.</param>
+        /// <returns>An object of type <typeparamref name="TJson"/>.</returns>
         public TJson Deserialize(string value)
         {
             return DeserializeObject<TJson>(value);
         }
 
         /// <summary>
-        /// DeserializeObject
+        /// Deserializes the specified JSON string to an object of type <typeparamref name="T"/>.
         /// </summary>
-        /// <typeparam name="T"></typeparam>
-        /// <param name="value"></param>
-        /// <returns></returns>
+        /// <typeparam name="T">The type to deserialize to.</typeparam>
+        /// <param name="value">The JSON string to deserialize.</param>
+        /// <returns>An object of type <typeparamref name="T"/>.</returns>
         public T DeserializeObject<T>(string value)
         {
             return JsonConvert.DeserializeObject<T>(value, settings);